requirements.txt
attrs==24.2.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.8
kaitaistruct==0.10
outcome==1.3.0.post0
pyasn1==0.6.0
pycparser==2.22
pydivert==2.1.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.24.0
setuptools==74.1.2
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
zstandard==0.23.0
